Extract ERC20 demo tests into its own file
